--- src/main/java/kr/hhplus/be/server/application/order/CancelOrderPaidEvent.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.application.order;
-
-import java.math.BigDecimal;
-
-public record CancelOrderPaidEvent(
-	Long orderId,
-	BigDecimal totalPrice
-) {
-}
