--- src/main/java/kr/hhplus/be/server/domain/dlqRecord/DLQRecordRepository.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.domain.dlqRecord;
-
-import java.util.List;
-
-public interface DLQRecordRepository{
-	List<DLQRecord> findAllByResolvedNot();
-	List<DLQRecord> findAllByEventType();
-	List<DLQRecord> findAllByUserId();
-	void save(DLQRecord dlqRecord);
-}
