+++ src/main/java/kr/hhplus/be/server/domain/order/event/CancelOrderPaidFailEvent.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.application.order;
+package kr.hhplus.be.server.domain.order.event;
 
 import java.math.BigDecimal;
 import java.util.HashMap;
